Restructure parser folder
