+++ src/platform/packages/shared/kbn-esql-ast/src/parser/README.md
@@ -21,13 +21,25 @@ The parser is structured as follows:
 ```
 src/
 |- parser/                            Contains the logic to parse the ES|QL query and generate the AST.
-|  |- factories.ts                    Contains AST node factories.
-|  |- antlr_error_listener.ts         Contains code which traverses ANTLR CST and collects syntax errors.
-|  |- esql_ast_builder_listener.ts    Contains code which traverses ANTLR CST and builds the AST.
-|
-|- antlr/                             Contains the autogenerated ES|QL ANTLR grammar files and various compilation assets.
-   |- esql_lexer.g4                   Contains the ES|QL ANTLR lexer grammar.
-   |- esql_parser.g4                  Contains the ES|QL ANTLR parser grammar.
+|  |- index.ts                        Main parser exports - primary entry point for consumers.
+|  |- core/                           High-level parsing logic and AST building.
+|  |  |- parser.ts                    Main Parser class with parse(), parseCommand(), parseExpression() methods.
+|  |  |- cst_to_ast_converter.ts      Converts ANTLR CST to ES|QL AST.
+|  |  |- esql_error_listener.ts       Collects syntax errors during parsing.
+|  |  |- constants.ts                 Parser constants and configuration.
+|  |  |- helpers.ts                   Utility functions for parsing operations.
+|  |  |- types.ts                     Parser-specific type definitions.
+|  |  └- index.ts                     Exports from the core parsing module.
+|  |
+|  |- antlr/                          ANTLR-generated grammar files and assets.
+|  |  |- esql_lexer.g4                ES|QL ANTLR lexer grammar.
+|  |  |- esql_parser.g4               ES|QL ANTLR parser grammar.
+|  |  |- esql_lexer.ts                Generated TypeScript lexer.
+|  |  |- esql_parser.ts               Generated TypeScript parser.
+|  |  |- promql_*.g4 / promql_*.ts    PromQL grammar and generated files.
+|  |  └- lexer_config.js              Lexer configuration for ANTLR.
+|  |
+|  └- __tests__/                      Parser tests and test utilities.
 ```
 
 ## Usage
