+++ src/platform/packages/shared/kbn-esql-ast/src/parser/__tests__/header.set.test.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { Parser } from '../parser';
+import { Parser } from '../core/parser';
 
 describe('SET instruction parsing', () => {
   describe('single SET instruction', () => {
