+++ src/platform/packages/shared/kbn-esql-ast/src/parser/core/constants.ts
