+++ src/platform/packages/shared/kbn-esql-ast/src/parser/core/esql_error_listener.ts
@@ -9,7 +9,7 @@
 
 import * as antlr4 from 'antlr4';
 import { getPosition } from './tokens';
-import type { EditorError } from '../types';
+import type { EditorError } from '../../types';
 
 // These will need to be manually updated whenever the relevant grammar changes.
 const SYNTAX_ERRORS_TO_IGNORE = [
