+++ src/platform/packages/shared/kbn-esql-ast/src/parser/core/helpers.ts
