+++ src/platform/packages/shared/kbn-esql-ast/src/parser/core/parser.ts
@@ -11,10 +11,10 @@ import { CharStreams, type Token } from 'antlr4';
 import { CommonTokenStream, type CharStream } from 'antlr4';
 import { ESQLErrorListener } from './esql_error_listener';
 import { attachDecorations, collectDecorations } from './decorations';
-import { Builder } from '../builder';
+import { Builder } from '../../builder';
 import { CstToAstConverter } from './cst_to_ast_converter';
-import { default as ESQLLexer } from '../antlr/esql_lexer';
-import { default as ESQLParser } from '../antlr/esql_parser';
+import { default as ESQLLexer } from '../../antlr/esql_lexer';
+import { default as ESQLParser } from '../../antlr/esql_parser';
 import type {
   ESQLAst,
   ESQLAstExpression,
@@ -24,9 +24,9 @@ import type {
   ESQLMap,
   ESQLProperNode,
   EditorError,
-} from '../types';
-import { isFunctionExpression, isProperNode, isQuery, isMap, isCommand } from '../ast/is';
-import { singleItems } from '../visitor/utils';
+} from '../../types';
+import { isFunctionExpression, isProperNode, isQuery, isMap, isCommand } from '../../ast/is';
+import { singleItems } from '../../visitor/utils';
 import { DEFAULT_CHANNEL, SOURCE_COMMANDS } from './constants';
 import type { EsqlParsingTarget } from './types';
