+++ src/platform/packages/shared/kbn-esql-ast/src/parser/core/tokens.ts
@@ -9,7 +9,7 @@
 
 import type { Token } from 'antlr4';
 import { DEFAULT_CHANNEL } from './constants';
-import type { ESQLLocation } from '../types';
+import type { ESQLLocation } from '../../types';
 
 /**
  * Finds all tokens in the given range using binary search. Allows to further
