+++ src/platform/packages/shared/kbn-esql-ast/src/parser/core/types.ts
@@ -7,9 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import type * as esql_parser from '../antlr/esql_parser';
+import type * as esql_parser from '../../antlr/esql_parser';
 import type { CstToAstConverter } from './cst_to_ast_converter';
-import type { ESQLAstComment } from '../types';
+import type { ESQLAstComment } from '../../types';
 
 /**
  * Lines of decorations per *whitespace line*. A *whitespace line* is a line
